+++ packages/atomic_widget/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.2
+- feat: Add `DeviceSize.webBrowser` and `DeviceSize.fullScreenWebBrowser`.
+- feat: Add iOS, Android, and Web platform lists.
+
 ## 0.1.1
 - chore: Upgrade `Dart` version to `3.2.4` and `Flutter` version to `3.16.7`.
 - chore: Upgrade dependent package versions.